action_service/requirements.txt
blinker==1.8.2
boto3==1.34.133
botocore==1.34.133
click==8.1.7
easydict==1.13
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sympy==1.12.1
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
imageio
imageio[ffmpeg]
